Allow editing of labels and style classes
For nodes and edges, allow the adding of labels and the addition or changing
of using one of the classes specified in styles.

Not as close to the whiteboard experience as I would like, but a temporary
solution that at least offers the functionality pending a more complicated
submittion that does it nicer.